Move marginCurrencyList sync-schema to separate file
